feat(api): update step resolver bridge composition
