+++ apps/api/src/.env.production
@@ -79,3 +79,5 @@ PLAIN_IDENTITY_VERIFICATION_SECRET_KEY='PLAIN_IDENTITY_VERIFICATION_SECRET_KEY'
 STEP_RESOLVER_CF_ACCOUNT_ID=
 STEP_RESOLVER_CF_API_TOKEN=
 STEP_RESOLVER_CF_DISPATCH_NAMESPACE=
+STEP_RESOLVER_DISPATCH_URL=
+STEP_RESOLVER_HMAC_SECRET=
