+++ apps/worker/src/app/workflow/workflow.module.ts
@@ -8,7 +8,7 @@ import {
   CompileTemplate,
   ConditionsFilter,
   CreateExecutionDetails,
-  ExecuteStepResolver,
+  ExecuteStepResolverRequest,
   GetDecryptedIntegrations,
   GetLayoutUseCase,
   GetNovuLayout,
@@ -191,7 +191,7 @@ const USE_CASES = [
   CompileInAppTemplate,
   InboundEmailParse,
   ExecuteBridgeJob,
-  ExecuteStepResolver,
+  ExecuteStepResolverRequest,
   GetPreferences,
   GetSubscriberSchedule,
   ResolveChannelEndpoints,
