+++ libs/application-generic/src/schemas/control/email-control.schema.ts
@@ -18,6 +18,7 @@ export const emailControlZodSchema = z
         name: z.string().optional(),
       })
       .optional(),
+    stepResolverHash: z.string().optional(),
   })
   .strict();
