+++ libs/application-generic/src/usecases/execute-bridge-request/index.ts
@@ -1,5 +1,8 @@
 export {
+  BridgeError,
   ExecuteBridgeRequestCommand,
   ExecuteBridgeRequestDto,
+  ProcessError,
 } from './execute-bridge-request.command';
 export { ExecuteBridgeRequest } from './execute-bridge-request.usecase';
+export { ExecuteFrameworkRequest } from './execute-framework-request.usecase';
